Fix usage of variables for fixed values
Some values which are fixed by the docker compose file should not be set
as variables, and hardcoded everywhere as required.

This also untangles the use of the variables DB_HOST and DB_PORT to be
only about the exposed hostname and port, whenever the DB engine is
exposed on the server interface.